Set gait_type to 'walk_like' when empty. Move PerformGaitAction to its own class.
